LayoutTests/accessibility/range-alter-by-percent-expected.txt

  
This tests that decrement and increment alter a range type input element by five percent or one (whichever is larger) when no step is specified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS largeRange.intValue is 50
PASS largeRange.intValue is 55
PASS largeRange.intValue is 60
PASS largeRange.intValue is 55
PASS largeRange.intValue is 50
PASS smallRange.intValue is 5
PASS smallRange.intValue is 6
PASS smallRange.intValue is 7
PASS smallRange.intValue is 6
PASS smallRange.intValue is 5
PASS successfullyParsed is true

TEST COMPLETE
